Fix ticket #4
The returned mutex was not held until the MutexLocker object is destructed.

It allowed another thread, the image acquisition thread, overwrites the
latest image to be displayed in the canvas.

Added a member variable to hold the locked mutex. Having this variable, the
locked mutex is held until the MutexLocker object is destructed.

Fixes ticket #4.