Create daemon threads if Harvester is in the interactive mode
If Harvester is running in the interactive mode, there’s no way to catch
the timing where quit() was called (then the SystemExit exception is
raised) because the responsibility is on the interactive interpreter side
such as Python or IPython. It means we can’t stop our threads by
ourselves. To avoid this situation happens, we create daemon threads
instead of non-daemon threads. This means we can’t explicitly be
responsible for cleaning up our responsible objects but at least we are
sure that our threads will never block the interpreter exits because
they're daemon threads.

This change may be relevant to the following issues:

https://github.com/genicam/harvesters/issues/40
https://github.com/genicam/harvesters/issues/47