Change the default value of sleep_duration
It's the same amount of time that Harvester GUI uses on PyQt and it
guarantees Harvester Core can acquire images at 1000 fps on Kazunari's
laptop. If you expect to acquire much more faster, call
create_image_acquirer(n, sleep_duration=0.0). However, you should note
that there could be a case where it blocks the image acquisition
thread in Harvester Core which drags the acquisition frame rate. In
that case, there'd be no way to get the expected frame rate and we have
not had a firm solution for that.