Strip leading slashes from local: xml url
Fixes: #126

According to the GenICam standard 1.4, the /// after local: should be a valid configuration.  But we need to strip it out so it doesn't affect saving the file on the local machine.

4.1.2.1 Module Register Map (Recommended)
A URL in the form “local:[///]filename.extension;address;length[?SchemaVersion=x.x.x]”
indicates that the XML description file is located in the module’s virtual register map. The
square brackets are optional.