bugfix: allow chunk data
There was a problem for some with fetching buffers containing chunk data. I believe this is the cause of issue #269 and #276. 

The problem appears to be the buffer containing padding after the end of the chunk data and tag. To solve the issue everything in the buffer after chunk payload size can be removed since chunk payload size should point to the very end of the chunk tag, which should be at the very end of the buffer.

self._chunk_adapter.attach_buffer(buffer.raw_buffer[:buffer.delivered_chunk_payload_size]) is enough to solve the issue, however the if statment should have better performance (if slicing the buffer creates a copy of it).